% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileHandlers.R
\name{ddsHandler}
\alias{ddsHandler}
\title{ddsHandler}
\usage{
ddsHandler(file, output = ".", title = "")
}
\arguments{
\item{file}{Filepath to a DESeqDataSet .rds file OR a DESeqDataSet object}

\item{output}{Path to output directory, default current directory}
}
\value{
a list specific to the type of object
}
\description{
ddsHandler
}
