% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileHandlers.R
\name{designHandler}
\alias{designHandler}
\title{tableHandler}
\usage{
designHandler(design_file, pattern = "design_")
}
\arguments{
\item{design_file}{a design file of either .csv or dataframe type}
}
\value{
a list with a samples table and a file prefix
}
\description{
tableHandler
}
