% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaHelper.R
\name{pcaHelper}
\alias{pcaHelper}
\title{pcaHelper}
\usage{
pcaHelper(
  dds,
  title = "",
  label = "",
  PC1 = 1,
  PC2 = 2,
  save = TRUE,
  output = ".",
  saveToDir = FALSE
)
}
\arguments{
\item{dds}{DESeqDataSet object or file path to dds RDS object}

\item{title}{Character, title of the plot}

\item{label}{Character, additional text, may be removed}

\item{PC1}{Integer, first principle component}

\item{PC2}{Integer, second principle component}

\item{save}{Logical, save to output directory if true}

\item{output}{Character, file path to output directory, default current directory}

\item{saveToDir}{Logical,}
}
\value{
a PCA plot
}
\description{
Create a PCA plot that can also adjust for batch effects.
}
