% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qBulkSeq-package.R
\docType{package}
\name{qBulkSeq-package}
\alias{qBulkSeq}
\alias{qBulkSeq-package}
\title{qBulkSeq: An R package to quickly analyze Bulk RNA-Seq data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Karthik Madhusudhanan \email{karthikmprojects@gmail.com} (\href{https://orcid.org/0009-0008-4504-2729}{ORCID})

}
\keyword{internal}
