% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickDESeq.R
\name{quickDESeq}
\alias{quickDESeq}
\title{Quickly extract a DESeqDataSet with a simple design}
\usage{
quickDESeq(
  design_file,
  pattern = "design_",
  output = ".",
  terms = c("condition"),
  alignType = "kallisto",
  save = TRUE,
  mincounts = 5,
  filter = 2
)
}
\arguments{
\item{design_file}{CSV file path or data.frame with sample metadata.}

\item{pattern}{Pattern to substitue out for the design file when extracting title, default "design_"}

\item{output}{File path to output directory}

\item{terms}{Character vector of terms for the model formula.}

\item{alignType}{What type of aligner was used to align BAM files, default "kallisto"}

\item{save}{Logical, whether to save the DESeq object as an RDS.}

\item{mincounts}{Integer, minimum counts filtering threshold}

\item{filter}{Integer, minimum count filter threshold.}
}
\value{
A DESeqDataSet object
}
\description{
Create a DESeqDataSet object using tximport and design terms from a design file.
}
\details{
IMPORTANT: Design file MUST be either a table or a .csv with a naming convention
of "design_\link{VARNAME}".

Additionally, design files should have the following columns:
\itemize{
\item sample: name of the sample (e.g. A5, treated, etc.)
\item file: name of the file (normally the same as the sample)
\item folder: folder where the sample's abundance.h5 file can be found
\item design terms: columns for desired terms (e.g. genotype, time, treatment, etc.)
\item batch: batch variable to account for (if you would like to explicitly remove the batch effect,
select batch = TRUE)
}
}
