% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickHeatmap.R
\name{quickHeatmap}
\alias{quickHeatmap}
\title{quickHeatmap}
\usage{
quickHeatmap(
  file = NULL,
  title = "",
  output = ".",
  targets = NULL,
  info = "condition",
  clustRows = TRUE,
  clustCols = TRUE,
  colOrder = NULL,
  labelRows = TRUE,
  numRows = 50,
  num = "",
  denom = "",
  padj = 0.05,
  save = TRUE,
  boldSigs = TRUE,
  onlySigs = FALSE,
  onlyNonSigs = FALSE,
  color = colorRampPalette(list.reverse(brewer.pal(11, "PRGn")))(100)
)
}
\arguments{
\item{file}{A DESeqDataSet or path to a DESeq .rds object}

\item{title}{Character, title of plot}

\item{output}{Character, path to output directory, default current directory}

\item{targets}{Character list, list of target genes to construct heatmap from, defaults to top 50 if empty}

\item{info}{Basis of differential expression analysis, defaults to "condition"}

\item{clustRows}{Logical, whether to cluster rows or not. default TRUE}

\item{clustCols}{Logical, whether to cluster columns or not, default TRUE}

\item{colOrder}{Character vector, order for columns}

\item{labelRows}{Logical, whether to label rows or not, default TRUE}

\item{numRows}{Integer, number of rows to display, default 50}

\item{num}{Character, the numerator of the comparison}

\item{denom}{Character, the denominator of the comparison}

\item{padj}{Double, value for p-adjusted value threshold, default 0.05}

\item{save}{Logical, whether to save to output directory or not, default TRUE}

\item{boldSigs}{Logical, if significant genes should be bolded, default TRUE}

\item{onlySigs}{Logical, if heatmap should only should significant genes, default FALSE}

\item{onlyNonSigs}{Logical, if heatmap should only should NON-significant genes, default FALSE}

\item{color}{A color palette, default colorRampPalette(list.reverse(brewer.pal(11, "PRGn")))(100)}
}
\value{
A heatmap
}
\description{
quickHeatmap
}
