% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickPCA.R
\name{quickPCA}
\alias{quickPCA}
\title{pcaHelper}
\usage{
quickPCA(
  dds,
  title = "",
  label = "",
  PC1 = 1,
  PC2 = 2,
  intgroup = "condition",
  blind = FALSE,
  batch = NULL,
  save = FALSE,
  output = ".",
  saveToDir = FALSE,
  width = 1600,
  height = 1200
)
}
\arguments{
\item{dds}{DESeqDataSet object or file path to dds RDS object}

\item{title}{Character, title of the plot}

\item{label}{Character, additional text, may be removed}

\item{PC1}{Integer, first principal component}

\item{PC2}{Integer, second principal component}

\item{intgroup}{Character or character vector indicating which variables in colData to use for grouping}

\item{batch}{Logical, apply batch correction using the 'year' column if TRUE}

\item{save}{Logical, save plot to output directory if TRUE}

\item{output}{Character, file path to output directory, default is current directory}

\item{saveToDir}{Logical, extract output directory and title from file path if TRUE}

\item{width}{Numeric, width of the saved plot in pixels}

\item{height}{Numeric, height of the saved plot in pixels}
}
\value{
A PCA ggplot object
}
\description{
Create a PCA plot that can also adjust for batch effects.
}
