% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickVolcano.R
\name{quickVolcano}
\alias{quickVolcano}
\title{quickResults}
\usage{
quickVolcano(
  results,
  title = "",
  output = "",
  targets = "",
  xlim = c(min(results[["log2FoldChange"]], na.rm = TRUE) - 1.5,
    max(results[["log2FoldChange"]], na.rm = TRUE) + 1.5),
  ylim = c(0, max(-log10(results[["padj"]]), na.rm = TRUE) + 5),
  numTopGenes = 10,
  showSigs = TRUE,
  labelSize = 4,
  save = FALSE,
  displayNumSigs = TRUE,
  FCcutoff = 1,
  pCutoff = 0.05
)
}
\arguments{
\item{results}{a table of results of either a tibble format or csv}

\item{title}{Character, the name of the plot to be shown}

\item{output}{Character, path to save image to, default current directory}

\item{save}{Logical, whether to save the plot}

\item{FCcutoff}{Double, cutoff for log2FoldChange threshold, default 1}

\item{pCutoff}{Double, cutoff for p-adjusted value for significance, default 0.05}
}
\value{
A volcano plot
}
\description{
quickResults
}
