% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileHandlers.R
\name{resultsHandler}
\alias{resultsHandler}
\title{resultsHandler}
\usage{
resultsHandler(results, title = "", output = ".")
}
\arguments{
\item{results}{a path to a results table in the form of "RES-name.csv" or a results table dataframe}

\item{title}{title of the results table, defaults to deparsed name}

\item{output}{path to save what this table is used for, defaults to current directory}
}
\value{
A list containing a results table, title, and output
}
\description{
resultsHandler
}
