% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileHandlers.R
\name{tableHandler}
\alias{tableHandler}
\title{tableHandler}
\usage{
tableHandler(design_file, pattern)
}
\arguments{
\item{design_file}{a design file of either .csv or dataframe type}
}
\value{
a list with a samples table and a file prefix
}
\description{
tableHandler
}
