% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickDESeq.R
\name{quickDESeq}
\alias{quickDESeq}
\title{quickDESeq}
\usage{
quickDESeq(
  design_file,
  pattern = "design_",
  output = ".",
  terms = c("condition"),
  alignType = "kallisto",
  csv = TRUE,
  save = TRUE,
  mincounts = 5,
  filter = 2
)
}
\arguments{
\item{design_file}{CSV file path or data.frame with sample metadata.}

\item{output}{File path to output directory}

\item{terms}{Character vector of terms for the model formula.}

\item{csv}{Logical, TRUE if design_file is a CSV.}

\item{save}{Logical, whether to save the DESeq object as an RDS.}

\item{mincounts}{Integer, minimum counts filtering threshold}

\item{filter}{Integer, minimum count filter threshold.}
}
\value{
A DESeqDataSet object
}
\description{
Create a DESeqDataSet object using tximport and design terms from a design file.
}
