% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickHeatmap.R
\name{quickHeatmap}
\alias{quickHeatmap}
\title{quickHeatmap}
\usage{
quickHeatmap(
  file = NULL,
  title = "",
  targets = NULL,
  info = "condition",
  clustRows = TRUE,
  clustCols = TRUE,
  labelRows = TRUE,
  n = 50,
  num = "dcas9_ANRIL",
  denom = "dcas9_NONE",
  color = colorRampPalette(list.reverse(brewer.pal(11, "PRGn")))(100)
)
}
\arguments{
\item{file}{A DESeqDataSet or path to a DESeq .rds object}

\item{title}{Character, title of plot}

\item{targets}{Character list, list of target genes to construct heatmap from, defaults to top 50 if empty}

\item{info}{Basis of differential expression analysis, defaults to "condition"}

\item{clustRows}{Logical, whether to cluster rows or not. default TRUE}

\item{clustCols}{Logical, whether to cluster columns or not, default TRUE}

\item{labelRows}{Logical, whether to label rows or not, default TRUE}

\item{n}{Integer, number of rows to display, default 50}

\item{num}{Character, the numerator of the comparison}

\item{denom}{Character, the denominator of the comparison}

\item{color}{A color palette, default colorRampPalette(list.reverse(brewer.pal(11, "PRGn")))(100)}
}
\value{
A heatmap
}
\description{
quickHeatmap
}
